+++ blockfrost-api/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version [0.7.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.6.0.0...v0.7.0.0) (2022-10-11)
+
+* Changes
+  * Added `ScriptHashList` newtype instead of overlapping instance for `[ScriptHash]` which is used
+    as a result of `_listScripts` in `ScriptsAPI`
+
 # Version [0.6.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.5.0.0...v0.6.0.0) (2022-08-31)
 
 * Additions