+++ blockfrost-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version [0.7.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.6.0.0...v0.7.0.0) (2022-10-11)
+
+* Changes
+  * `listScript` and `listScripts'` now return `ScriptHashList` newtype
+    instead of `[ScriptHash]` due to overlapping instance conflicts
+
 # Version [0.6.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/v0.5.0.0...v0.6.0.0) (2022-08-31)
 
 * Additions